Add Character marker component and change player_id to u8
Introduce Character as a marker to distinguish player-controlled entities
from other Player-tagged objects. Change player_id from usize to u8 across
all components and messages, and update BeamChargesChanged to carry player_id
directly instead of an Entity handle.

+++ src/components/markers.rs
@@ -12,7 +12,7 @@ pub struct Ground;
 #[derive(Component, Reflect, Default)]
 #[reflect(Component, Default)]
 pub struct HPBar {
-    pub player_id: usize,
+    pub player_id: u8,
 }
 
 #[derive(Component, Default)]
